get function's fParam (in FunctionValue)

+++ src/main/java/cn/edu/nju/software/ir/value/ConstValue.java
@@ -55,7 +55,7 @@ public class ConstValue extends ValueRef {
         } else if (type instanceof BoolType) {
             return value.toString();
         } else if (type instanceof FloatType) {
-            long floatBits = Double.doubleToRawLongBits(Double.parseDouble(value.toString()));
+            long floatBits = Double.doubleToRawLongBits(Float.parseFloat(value.toString()));
             String hex = Long.toHexString(floatBits);
             return "0x" + hex;
         } else {
